refactor: made services file an example only
